listingservice/requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.5.0
tzdata==2024.1
pika==1.2.0
django-cors-headers~=4.1.0